fix(cron): extract reply skip-reason leaf to break import cycle and refresh prompt snapshots
