+++ test/fixtures/agents/prompt-snapshots/codex-runtime-happy-path/telegram-heartbeat-codex-tool.md.diff
@@ -1,5 +1,5 @@
---- telegram-direct-codex-message-tool.md	sha256=3a723ba690b1938399f61d2dde771a424a77a34ae2aee1c3c567de3d3100ec5f
-+++ telegram-heartbeat-codex-tool.md	sha256=6727aba7851d91878e03b76e03fc9642f88eca88f4ee5859ba75f47046dbeea3
+--- telegram-direct-codex-message-tool.md	sha256=5b4b164ad416cb871634c615e6e1abfbfa509da37945ea1215e43d4f6998029b
++++ telegram-heartbeat-codex-tool.md	sha256=1d575d3564e0435c836b6396cc7aa98a3ffdfc0e91ab479f1eb50a62c3d4b303
 @@ -1,1 +1,1 @@
 -# Telegram Direct Codex Message Tool Turn
 +# Telegram Direct Codex Heartbeat Tool Turn
@@ -32,20 +32,20 @@
 -    "dynamicToolsFrom": "codex-dynamic-tools.telegram-direct.json",
 +    "dynamicToolsFrom": "codex-dynamic-tools.heartbeat-turn.json",
 @@ -235,2 +230,2 @@
--    "chars": 54657,
--    "roughTokens": 13665
-+    "chars": 56150,
-+    "roughTokens": 14038
+-    "chars": 55447,
+-    "roughTokens": 13862
++    "chars": 56940,
++    "roughTokens": 14235
 @@ -243,2 +238,2 @@
 -    "chars": 27464,
 -    "roughTokens": 6866
 +    "chars": 27806,
 +    "roughTokens": 6952
 @@ -247,2 +242,2 @@
--    "chars": 82123,
--    "roughTokens": 20531
-+    "chars": 83958,
-+    "roughTokens": 20990
+-    "chars": 82913,
+-    "roughTokens": 20729
++    "chars": 84748,
++    "roughTokens": 21187
 @@ -251,2 +246,2 @@
 -    "chars": 863,
 -    "roughTokens": 216
